backend/requirements.txt
# AI & LLM
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
openai>=1.40.0

# Microsoft 365
msgraph-sdk>=1.0.0
azure-identity>=1.17.0
msal>=1.31.0

# Google Services
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
google-generativeai>=0.8.0          # Gemini AI
langchain-google-genai>=2.0.0       # LangChain + Gemini

# Vector Database & RAG
chromadb>=0.5.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.32.0

# Scheduler & Cache
apscheduler>=3.10.0
redis>=5.0.0

# Google Calendar
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pytz>=2024.1

# UI
streamlit>=1.38.0
fastapi>=0.115.0
uvicorn>=0.32.0

# Auth & API
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# Utils
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
loguru>=0.7.0
python-dateutil>=2.9.0
